backend_2/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pyautogen==0.2.3
pydantic==2.4.2
websockets==11.0.3
python-dotenv==1.0.0
uuid==1.30